+++ tests/mat3.rs
@@ -2,7 +2,7 @@
 mod support;
 
 macro_rules! impl_mat3_tests {
-    ($t:ident, $const_new:ident, $newmat3:ident, $mat3:ident, $mat4:ident, $quat:ident, $newvec3:ident, $vec3:ident, $vec2:ident) => {
+    ($t:ident, $const_new:ident, $newmat3:ident, $mat3:ident, $mat2:ident, $mat4:ident, $quat:ident, $newvec3:ident, $vec3:ident, $vec2:ident) => {
         use core::$t::INFINITY;
         use core::$t::NAN;
         use core::$t::NEG_INFINITY;
@@ -212,6 +212,15 @@ macro_rules! impl_mat3_tests {
         }
 
         #[test]
+        fn test_from_mat2() {
+            let m2 = $mat2::from_cols_array_2d(&[[1.0, 2.0], [3.0, 4.0]]);
+            let m3 = $mat3::from_mat2(m2);
+            assert_eq!(
+                $mat3::from_cols_array_2d(&[[1.0, 2.0, 0.0], [3.0, 4.0, 0.0], [0.0, 0.0, 1.0]]),
+                m3
+            );
+        }
+        #[test]
         fn test_from_mat4() {
             let m4 = $mat4::from_cols_array_2d(&[
                 [1.0, 2.0, 3.0, 4.0],
@@ -219,7 +228,7 @@ macro_rules! impl_mat3_tests {
                 [9.0, 10.0, 11.0, 12.0],
                 [13.0, 14.0, 15.0, 16.0],
             ]);
-            let m3 = $mat3::from(m4);
+            let m3 = $mat3::from_mat4(m4);
             assert_eq!(
                 $mat3::from_cols_array_2d(&[[1.0, 2.0, 3.0], [5.0, 6.0, 7.0], [9.0, 10.0, 11.0]]),
                 m3
@@ -361,7 +370,9 @@ macro_rules! impl_mat3_tests {
 
 mod mat3 {
     use super::support::deg;
-    use glam::{const_mat3, mat3, swizzles::*, vec3, vec3a, Mat3, Mat4, Quat, Vec2, Vec3, Vec3A};
+    use glam::{
+        const_mat3, mat3, swizzles::*, vec3, vec3a, Mat2, Mat3, Mat4, Quat, Vec2, Vec3, Vec3A,
+    };
 
     #[test]
     fn test_align() {
@@ -390,12 +401,14 @@ mod mat3 {
         );
     }
 
-    impl_mat3_tests!(f32, const_mat3, mat3, Mat3, Mat4, Quat, vec3, Vec3, Vec2);
+    impl_mat3_tests!(f32, const_mat3, mat3, Mat3, Mat2, Mat4, Quat, vec3, Vec3, Vec2);
 }
 
 mod mat3a {
     use super::support::deg;
-    use glam::{const_mat3a, mat3a, swizzles::*, vec3a, Mat3A, Mat4, Quat, Vec2, Vec3, Vec3A};
+    use glam::{
+        const_mat3a, mat3a, swizzles::*, vec3a, Mat2, Mat3A, Mat4, Quat, Vec2, Vec3, Vec3A,
+    };
 
     #[test]
     fn test_align() {
@@ -425,6 +438,7 @@ mod mat3a {
         const_mat3a,
         mat3a,
         Mat3A,
+        Mat2,
         Mat4,
         Quat,
         vec3a,
@@ -435,7 +449,7 @@ mod mat3a {
 
 mod dmat3 {
     use super::support::deg;
-    use glam::{const_dmat3, dmat3, dvec3, swizzles::*, DMat3, DMat4, DQuat, DVec2, DVec3};
+    use glam::{const_dmat3, dmat3, dvec3, swizzles::*, DMat2, DMat3, DMat4, DQuat, DVec2, DVec3};
 
     #[test]
     fn test_align() {
@@ -449,6 +463,7 @@ mod dmat3 {
         const_dmat3,
         dmat3,
         DMat3,
+        DMat2,
         DMat4,
         DQuat,
         dvec3,